Split Workspace core package
