+++ docs/en/architecture/extension-boundaries.md
@@ -140,6 +140,24 @@ such as `desc_mode` should be typed as `Literal["append", "override",
 "default"]`, while runtime validation should remain for untyped or dynamic
 callers.
 
+### Module Organization
+
+Core and builtins capabilities can be implemented as subdirectory packages.
+Prefer that shape when a feature has multiple architectural roles, such as a
+facade, manager, default implementation, registry, adapter, policy, or
+validation layer.
+
+Do not default new core or builtins work to a single file. First judge the
+feature's expected submodule volume and ownership boundaries. A single file is
+appropriate only when the capability is genuinely small and splitting it would
+be over-design.
+
+Landed examples include `core/Action`, `core/TriggerFlow`,
+`core/TaskDAGExecutor`, `core/Workspace`,
+`builtins/plugins/ExecutionEnvironmentProvider`, and
+`builtins/plugins/SkillsExecutor`. Keep public imports stable through package
+`__init__.py` files and top-level re-exports.
+
 ## Action And Execution Environment
 
 Action and Execution Environment are separate layers.
